% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_finaloutput.R
\name{adjust_DALYs}
\alias{adjust_DALYs}
\title{adjust_DALYs}
\usage{
adjust_DALYs(
  df,
  use_stochastic = F,
  cases_multiplier = 1,
  deaths_multiplier = 1
)
}
\arguments{
\item{df}{the final model output file}

\item{use_stochastic}{whether to adjust the model output in a stochastic way}

\item{cases_multiplier}{the single multiplier for cases, the default is 1 (un-adjusted)}

\item{deaths_multiplier}{the single multiplier for deaths, the default is 1 (un-adjusted)}
}
\value{
a dataframe with adjusted DALYs(YLLs) (same format as the final model output file)
}
\description{
adjust the DALYs(YLLs) in the final model output file
}
