% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_finaloutput.R
\name{adjust_deaths}
\alias{adjust_deaths}
\title{adjust_deaths}
\usage{
adjust_deaths(df, use_stochastic = F, multiplier = 1)
}
\arguments{
\item{df}{the final model output file}

\item{use_stochastic}{whether to adjust the model output in a stochastic way}

\item{multiplier}{the single multiplier for deaths, the default is 1 (un-adjusted)}
}
\value{
a dataframe with adjusted deaths (same format as the final model output file)
}
\description{
adjust the deaths in the final model output file
}
