% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{adjusted_montagu_coverage}
\alias{adjusted_montagu_coverage}
\title{adjusted_montagu_coverage}
\usage{
adjusted_montagu_coverage(coverage_sheet, cntrycode)
}
\arguments{
\item{coverage_sheet}{the montagu coverage csv file}

\item{cntrycode}{the country code}
}
\value{
a dataframe with the coverage for one-dose and two-dose for the specified country each year adjusted using elizabeth's formula
}
\description{
adjusted_montagu_coverage
}
