% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rasters.R
\name{align_rasters}
\alias{align_rasters}
\title{align_rasters}
\usage{
align_rasters(datapath, country, orig_raster)
}
\arguments{
\item{datapath}{path to input data}

\item{country}{country code}

\item{orig_raster}{Raster* object that should be aligned to that in WorldPop and GADM}
}
\value{
Raster* object with aligned extent and resolution at country level
}
\description{
Align a raster to the extent and resolution of the worldpop rasters and GADM shapefiles
}
