% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allocate_vaccine}
\alias{allocate_vaccine}
\title{allocate_vaccine}
\usage{
allocate_vaccine(datapath, modelpath, country, scenario, ...)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{...}{Optional parameters to pass to [`assign_vaccine_targets()`]. See [`assign_vaccine_targets()`] for defaults.}
}
\value{
dataframe with proportion of total population allocated with vaccines in admin units in a given country and year
}
\description{
Create dataframe of total population vaccination coverage according to the admin-level assignments from assign_vaccine_targets
}
