% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{alpha_table}
\alias{alpha_table}
\title{alpha_table}
\usage{
alpha_table(cache)
}
\arguments{
\item{cache}{where the filenames and files are saved}
}
\value{
alpha table
}
\description{
make a table of alpha
}
