% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_targeting.R
\name{assign_vaccine_targets}
\alias{assign_vaccine_targets}
\title{assign_vaccine_targets}
\usage{
assign_vaccine_targets(
  datapath,
  modelpath,
  country,
  scenario,
  targeting_strat = "incidence",
  campaign_cov = 0.8,
  num_skip_years = 0
)
}
\arguments{
\item{datapath}{path to external data}

\item{modelpath}{path to model input (montagu) data}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{targeting_strat}{Unique string that identifies how targets should be ranked (default = incidence)}

\item{campaign_cov}{Proportion of the target population vaccinated by the campaign (default = 0.8)}

\item{num_skip_years}{Skip targeting in locations vaccinated within the last X years (default = 0 --> no skipping)}
}
\description{
Assign vaccine to a country and year
}
