% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_case_by_district.R
\name{case_by_district}
\alias{case_by_district}
\title{case_by_district}
\usage{
case_by_district(case_raster_multilayer, admin_shp)
}
\arguments{
\item{case_raster_multilayer}{the multi-layer expected case raster}

\item{admin_shp}{the shapefiles for a certain admin level}
}
\description{
sum number of cases by each district
}
