% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{check_model_setting}
\alias{check_model_setting}
\title{check_model_setting}
\usage{
check_model_setting(
  working_dir = NULL,
  configpath = "configs/202110gavi-3",
  countries,
  scenario,
  surveillance_scenario,
  thresholds,
  nsamples,
  ...
)
}
\arguments{
\item{configpath}{path of configuration files}

\item{countries}{vector of all the countries included the diagnostic report}

\item{scenario}{takes value of either "campaign-default" or "no-vaccination"}

\item{surveillance_scenario}{scale of surveillance, takes "no-estimate", "global-estimate", or "district-estimate"}

\item{rawoutpath}{(might not need this)}

\item{threshold}{threshold chosen for the diagnostic report, takes one of 1/1000, 1/5000, or 1/10,000}
}
\value{
table with model settings & scenario parameters
}
\description{
Table of model settings & assumptions (ie, which scenario does the diagnostic report represent?)
}
