% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{check_outputs_availability}
\alias{check_outputs_availability}
\title{check_outputs_availability}
\usage{
check_outputs_availability(
  working_dir = NULL,
  rawoutpath = "output_raw/202110gavi-3",
  configpath = "configs/202110gavi-3",
  countries,
  scenario,
  surveillance_scenario,
  threshold,
  admin_level,
  ...
)
}
\arguments{
\item{rawoutpath}{path of output_raw in which all the model output files are stored}

\item{configpath}{path of configuration files}

\item{countries}{vector of all the countries included the diagnostic report}

\item{scenario}{takes value of either "campaign-default" or "no-vaccination"}

\item{surveillance_scenario}{scale of surveillance, takes "no-estimate", "global-estimate", or "district-estimate"}

\item{threshold}{threshold chosen for the diagnostic report, takes one of 1/1000, 1/5000, or 1/10,000}

\item{admin_level}{the admin level to check for output}
}
\value{
TRUE or FALSE indicating whether the output files for the specified countries and scenarios exist in output folder, and also return warning messages to check output of which country is missing
}
\description{
function to check if all the output files are there
}
