% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_skip_countries.R
\name{check_table_screening}
\alias{check_table_screening}
\title{check_table_screening}
\usage{
check_table_screening(spathname, country, scenario, threshold, vac_admin_level)
}
\arguments{
\item{spathname}{path to input incidence data}

\item{country}{country code}

\item{threshold}{incidence rate threshold}

\item{vac_admin_level}{the admin level to check}
}
\description{
check_table_screening
}
