% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R,
%   R/utils_process_rawoutput.R
\name{combine_eff_table_allISOs}
\alias{combine_eff_table_allISOs}
\title{combine_eff_table_allISOs}
\usage{
combine_eff_table_allISOs(df_target, admin_level)

combine_eff_table_allISOs(df_target, admin_level)
}
\arguments{
\item{df_target}{target_table generated and saved in cache}

\item{admin_level}{admin level(s) selected for the diagnostic report}

\item{countries}{a vector contains all the countries included in the diagnostic report}
}
\value{
a table summarizing target population and ocv efficiency for one or more countries (at country level)

a table summarizing target population and ocv efficiency for one or more countries (at country level)
}
\description{
combine tables from different admin level(s)

combine tables from different admin level(s)
}
