% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{combine_output}
\alias{combine_output}
\title{combine_output}
\usage{
combine_output(
  cache,
  surveillance_project_directory,
  output_to_combine,
  save_final_output = TRUE
)
}
\arguments{
\item{cache}{the cached environment}

\item{surveillance_project_directory}{surveillance project directory}

\item{output_to_combine}{output to combine}

\item{save_final_output}{whether or not to save the combined results in the final output folder}
}
\value{
cached combined results
}
\description{
combine the output from different countries
}
