% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_model_pop_raster}
\alias{create_model_pop_raster}
\title{create_model_pop_raster}
\usage{
create_model_pop_raster(datapath, modelpath, country, year, cache)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{year}{population year}

\item{cache}{montagu cache}
}
\value{
raster of model population
}
\description{
Generate a population raster from relative worldpop weights and standardized country population from Montagu
}
