% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_relative_worldpop_weights}
\alias{create_relative_worldpop_weights}
\title{create_relative_worldpop_weights}
\usage{
create_relative_worldpop_weights(datapath, country)
}
\arguments{
\item{datapath}{path to data}

\item{country}{country code}
}
\value{
raster of relative population weights according to worldpop
}
\description{
Generate a raster with the relative population weight of each cell, which will be used to extrapolate rasterized populations using the montagu population data
}
