% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_static_modelInputs.R
\name{create_static_modelInputs}
\alias{create_static_modelInputs}
\title{create_static_modelInputs}
\usage{
create_static_modelInputs(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  vacc_alloc,
  clean
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{vacc_alloc}{object returned from [`allocate_vaccine()`]}

\item{clean}{logical that indicates whether existing model output files (vacc & pop) should be deleted}
}
\value{
rasterStack of total population for model years
}
\description{
Create proportion of population vaccinated and total population rasterStacks. Write vaccination raster to file and export population raster
}
