% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_central_template.R
\name{export_central_template}
\alias{export_central_template}
\title{export_central_template}
\usage{
export_central_template(stoch_output)
}
\arguments{
\item{stoch_output}{stochastic outputs, as from `[export_country_stoch_template()]`}
}
\value{
dataframe
}
\description{
Export a central burden template file (mean) using any stochastic template file output
}
