% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_country_stoch_template.R
\name{export_country_stoch_template}
\alias{export_country_stoch_template}
\title{export_country_stoch_template}
\usage{
export_country_stoch_template(
  modelpath,
  country,
  scenario,
  rawoutpath,
  outpath
)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{outpath}{path to final model output files}
}
\value{
dataframe
}
\description{
Export a stochastic burden template file for a single country
}
