% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_aoi}
\alias{generate_aoi}
\title{generate_aoi}
\usage{
generate_aoi(country)
}
\arguments{
\item{country}{country code}
}
\value{
numeric value of cfr (deaths/cases)
}
\description{
Generate average age of infection
}
