% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_indirect_incidence_mult}
\alias{generate_indirect_incidence_mult}
\title{generate_indirect_incidence_mult}
\usage{
generate_indirect_incidence_mult()
}
\value{

}
\description{
Using studies on indirect vaccine protection from Kolkota and Matlab, generate a function that takes the level of vaccination coverage and returns a multiplier indicating the percentage reduction in incidence due to indirect vaccine protection. These represent the baseline parameters for indirect vaccine protection.
}
