% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_infectionDuration}
\alias{generate_infectionDuration}
\title{generate_infectionDuration}
\usage{
generate_infectionDuration()
}
\value{
numeric value of infection duration
}
\description{
Generate infection duration in years
}
