% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pct_protect_function}
\alias{generate_pct_protect_function}
\title{generate_pct_protect_function}
\usage{
generate_pct_protect_function(my_trunc_year = 5, my_ve_scen = "base")
}
\arguments{
\item{my_ve_scen}{}

\item{my_trun_year}{}
}
\value{
dataframe with estimates for unweighted and weighted mean vaccination campaign coverage proportion across coverage surveys in the review
}
\description{
Using vaccine efficacy studies in Bi et al. (2017), generate a function that takes the year since vaccination and provides an estimate of the direct ve. Vaccine efficacy declines to 0 after my_trunc_year years
}
