% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pct_protect_function_one_dose_under5}
\alias{generate_pct_protect_function_one_dose_under5}
\title{generate_pct_protect_function_one_dose_under5}
\usage{
generate_pct_protect_function_one_dose_under5(
  my_trunc_year = 0,
  my_ve_scen = "base",
  study_type = "all"
)
}
\arguments{
\item{my_trunc_year}{}

\item{my_ve_scen}{}

\item{study_type}{}
}
\value{
dataframe with estimates for unweighted and weighted mean vaccination campaign coverage proportion across coverage surveys in the review
}
\description{
Using vaccine efficacy studies in the new Hanmeng and Andrew study, generate a function that takes the year since vaccination and provides an estimate of the direct ve. Vaccine efficacy declines to 0 after my_trunc_year years
}
