% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pct_protect_function_two_dose}
\alias{generate_pct_protect_function_two_dose}
\title{generate_pct_protect_function_two_dose}
\usage{
generate_pct_protect_function_two_dose(
  my_trunc_year = 5,
  my_ve_scen = "base",
  study_type = "efficacy"
)
}
\arguments{
\item{my_ve_scen}{}

\item{my_trun_year}{}
}
\value{
dataframe with estimates for unweighted and weighted mean vaccination campaign coverage proportion across coverage surveys in the review
}
\description{
Using vaccine efficacy studies in the new Hanmeng and Andrew study, generate a function that takes the year since vaccination and provides an estimate of the direct ve. Vaccine efficacy declines to 0 after my_trunc_year years
}
