% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_targeting.R
\name{get_admin_population}
\alias{get_admin_population}
\title{get_admin_population}
\usage{
get_admin_population(pop, shp)
}
\arguments{
\item{pop}{population raster}

\item{shp}{admin-level sf object}
}
\value{
vector of populations by admin2 in order of admin2 names in the shapefile
}
\description{
Wrapper for exact_extract with a given population raster and shapefile to get admin-level worldpop population summaries for a given country. Note that the default settings use na.rm = TRUE
}
