% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{get_confirmed_incidence_rate}
\alias{get_confirmed_incidence_rate}
\title{get_confirmed_incidence_rate}
\usage{
get_confirmed_incidence_rate(
  rc_list,
  model_year,
  surveillance_scenario,
  omicron_dataset,
  rcs,
  use_mean_ir,
  mean_ir_span,
  testing_sensitivity
)
}
\arguments{
\item{use_mean_ir}{}
}
\description{
to get the confirmed incidence rate in the rc_list
}
