% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{get_filenames}
\alias{get_filenames}
\title{get_filenames}
\usage{
get_filenames(
  cache,
  surveillance_project_directory,
  pre_country = NULL,
  pre_vac_incid_thresholds = NULL,
  pre_admin_levels = NULL,
  no_vaccination_surveillance_scenario,
  no_vaccination_threshold
)
}
\arguments{
\item{cache}{the cached environment}

\item{surveillance_project_directory}{surveillance project directory}

\item{pre_country}{countries specified in the parameters}

\item{pre_vac_incid_thresholds}{incidence rate threshold specified in the parameters}

\item{pre_admin_levels}{admin levels specified in the parameters}

\item{no_vaccination_surveillance_scenario}{the surveillance scenario specified for the no vaccination simulation}
}
\value{
cached file names
}
\description{
plot the polygon with modeled cases
}
