% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{get_inc_tp_doses}
\alias{get_inc_tp_doses}
\title{get_inc_tp_doses}
\usage{
get_inc_tp_doses(
  working_dir,
  rawoutpath = "output_raw/202110gavi-3",
  configpath = "configs/202110gavi-3",
  countries,
  scenario,
  surveillance_scenario,
  threshold,
  admin_level,
  ...
)
}
\arguments{
\item{rawoutpath}{path of output_raw in which all the model output files are stored}

\item{configpath}{path of configuration files}

\item{countries}{vector of all the countries included the diagnostic report}

\item{scenario}{takes value of either "campaign-default" or "no-vaccination"}

\item{surveillance_scenario}{scale of surveillance, takes "no-estimate", "global-estimate", or "district-estimate"}

\item{threshold}{threshold chosen for the diagnostic report, takes one of 1/1000, 1/5000, or 1/10,000}
}
\description{
function to get incidence, target population and number of doses administered under selected campaign scenario, surveillance scenario and threshold
}
