% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{get_inc_tp_doses_helper}
\alias{get_inc_tp_doses_helper}
\title{get_inc_tp_doses_helper}
\usage{
get_inc_tp_doses_helper(rc, sum_level, vax_cov = 0.68, surveillance_scenario)
}
\arguments{
\item{rc}{a data frame of one scenario of one country}

\item{sum_level}{takes either "country" or "admin". If takes "admin", returns a table by admin units; if takes "country", returns a table for the whole country aggregating all admin units.}

\item{vax_cov}{vaccine coverage, default value is 0.68}

\item{surveillance_scenario}{scale of surveillance scenario, which will be used to calculate observed incidence}
}
\value{
a table with incidence, target population and number of doses administered, by year and cumulatively
}
\description{
Function to assist calculation of get_inc_tp_doses(), read in only one data frame a time
}
