% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_years}
\alias{get_model_years}
\title{get_model_years}
\usage{
get_model_years(modelpath, country, vacc_alloc, cache)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{vacc_alloc}{object returned from [`allocate_vaccine()`]}

\item{cache}{montagu cache}
}
\value{
list with model_years, start_year, and real_model_years
}
\description{
Create list of years where vaccine dynamics will be important
}
