% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pop_proportion_ocv1}
\alias{get_pop_proportion_ocv1}
\title{get_pop_proportion_ocv1}
\usage{
get_pop_proportion_ocv1(vacc_alloc, year)
}
\arguments{
\item{vacc_alloc}{the output of allocate_vaccine}

\item{year}{the vaccination year}
}
\value{
the proportion of vaccinated people that received one dose of OCV in that year or the last year with vaccination
}
\description{
function that gets vacc_alloc and year as inputs and returns the proportion of vaccinated people that got
one dose of the vaccine
}
