% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_singular_estimate.R
\name{get_singular_estimate}
\alias{get_singular_estimate}
\title{get_singular_estimate}
\usage{
get_singular_estimate(datapath, country, nyears = 5)
}
\arguments{
\item{datapath}{path to data}

\item{country}{country code}

\item{nyears}{numeric, number of years we want to calculate average annual #cases from (default = 5 and always most recent)}
}
\value{
an updated .csv file
}
\description{
Generate a dataset with singular estimate for non-raster countries
}
