% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_rawoutput.R
\name{get_tp_eff_median}
\alias{get_tp_eff_median}
\alias{get_dose_eff_median}
\title{get_tp_eff_median}
\usage{
get_dose_eff_median(rc)
}
\arguments{
\item{rc}{target table (can be target table of one or multiple countries)}
}
\description{
summarize median targeted pop and median efficiency of each country
}
