% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_age_targeted_vac_pro.R
\name{import_age_targeted_vac_pro}
\alias{import_age_targeted_vac_pro}
\title{import_age_targeted_vac_pro}
\usage{
import_age_targeted_vac_pro(
  modelpath,
  country,
  ve_direct = generate_pct_protect_function()
)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{ve_direct}{vaccine effect function (default: generate_pct_protect_function())}
}
\value{

}
\description{
fix the age proportion according to the special coverage
}
