% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_country_agePop}
\alias{import_country_agePop}
\title{import_country_agePop}
\usage{
import_country_agePop(modelpath, country, redownload = TRUE)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{redownload}{whether to redownload the file}
}
\value{
dataframe
}
\description{
Get age-specific (1-year) country population from standardized source in Montagu. [may be used in future for case-based targeting]
}
