% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_country_lifeExpectancy_1yr}
\alias{import_country_lifeExpectancy_1yr}
\title{import_country_lifeExpectancy_1yr}
\usage{
import_country_lifeExpectancy_1yr(modelpath, country, year)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{year}{year for which life expectancy should be returned}
}
\description{
Get total life expectancy at birth for a single year from standardized source in Montagu. [may be used in future for case-based targeting]
}
