% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_country_population_1yr}
\alias{import_country_population_1yr}
\title{import_country_population_1yr}
\usage{
import_country_population_1yr(modelpath, country, year, cache)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{year}{population year}

\item{cache}{montagu cache}
}
\value{
numeric value
}
\description{
Get total country population from standardized source in Montagu for a single year
}
