% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_country_proportion_under5}
\alias{import_country_proportion_under5}
\title{import_country_proportion_under5}
\usage{
import_country_proportion_under5(modelpath, country, year, redownload = TRUE)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{year}{year (year in which vaccination campaign takes place)}

\item{redownload}{whether to redownload the file}
}
\value{
the proportion of the population aged <5 per year for the specified country and year
}
\description{
Get proportion of country population aged 1-4 from standardized source in Montagu
}
