% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_coverage_scenario}
\alias{import_coverage_scenario}
\title{import_coverage_scenario}
\usage{
import_coverage_scenario(
  modelpath,
  country,
  scenario,
  num_doses = NULL,
  filter0 = FALSE,
  redownload = TRUE
)
}
\arguments{
\item{modelpath}{Path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{filter0}{logical for whether 0 vaccination years should be filtered out (default = FALSE)}

\item{redownload}{whether to redownload the file}
}
\value{
Dataframe with vaccination coverage for a single scenario and country. Years without vaccination are excluded from the returned dataframe. Countries without vaccination in any year return a null dataframe.
}
\description{
Imports a CSV file with vaccination coverage values for the scenario
}
