% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_coverage_scenario_custom}
\alias{import_coverage_scenario_custom}
\title{import_coverage_scenario_custom}
\usage{
import_coverage_scenario_custom(
  datapath,
  country,
  scenario,
  cache,
  filter0 = FALSE
)
}
\arguments{
\item{datapath}{Path to input files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{cache}{montagu cache object}

\item{filter0}{logical for whether 0 vaccination years should be filtered out (default = FALSE)}
}
\value{
Dataframe with vaccination coverage for a single scenario and country. Years without vaccination are excluded from the returned dataframe. Countries without vaccination in any year return a null dataframe.
}
\description{
Imports a CSV file with vaccination coverage values for the DRC Case study
}
