% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_disability_weight}
\alias{import_disability_weight}
\title{import_disability_weight}
\usage{
import_disability_weight()
}
\value{
numeric value for disability weight
}
\description{
Returns mean disability weight for severe diarrheal disease from IHME Global Burden of Disease 2016 and 2019 (same estimate)
}
