% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_int_country_population}
\alias{import_int_country_population}
\title{import_int_country_population}
\usage{
import_int_country_population(modelpath, country, cache, redownload = TRUE)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{cache}{montagu cache}

\item{redownload}{whether to redownload the file}
}
\value{
dataframe
}
\description{
Get 1-year country population from standardized source in Montagu. [may be used in future for case-based targeting]
}
