% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_templateFilename_prefix}
\alias{import_templateFilename_prefix}
\title{import_templateFilename_prefix}
\usage{
import_templateFilename_prefix(type, modelpath)
}
\arguments{
\item{type}{"stochastic", "central", "parameter"}

\item{modelpath}{path to Montagu files}
}
\value{
template filename prefix
}
\description{
Return string from filename template
}
