% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_custom_shapefile_by_country}
\alias{load_custom_shapefile_by_country}
\title{load_custom_shapefile_by_country}
\usage{
load_custom_shapefile_by_country(admin0 = FALSE)
}
\arguments{
\item{admin0}{whether to return an admin0 level sf (default = FALSE)}
}
\value{
shapefile in sf format
}
\description{
Load a custom shapefile for DRC health zones (for the DRC case study) or a generated admin0 level shapefile for a country and return sf object
}
