% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shapefile_by_country.R
\name{load_shapefile_by_country}
\alias{load_shapefile_by_country}
\title{load_shapefile_by_country}
\usage{
load_shapefile_by_country(datapath, country, simple = FALSE)
}
\arguments{
\item{datapath}{path to data}

\item{country}{country code}

\item{simple}{logical indicating whether simple (full country) shapefile should be used (default: FALSE)}
}
\value{
shapefile in sf format
}
\description{
Download admin2 or admin0 (simple) level shapefile for a country from GADM, if not already in shapefiles folder, and return sf object
}
