% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{load_shp0_by_country}
\alias{load_shp0_by_country}
\title{load_shp0_by_country}
\usage{
load_shp0_by_country(datapath, country)
}
\arguments{
\item{datapath}{path to input data}

\item{country}{country code}
}
\value{
shapefile
}
\description{
load_shp0_by_country
}
