% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{load_shp2_by_country}
\alias{load_shp2_by_country}
\title{load_shp2_by_country}
\usage{
load_shp2_by_country(datapath, country, simple = FALSE)
}
\arguments{
\item{datapath}{path to input data}

\item{country}{country code}

\item{simple}{return country level shapefile is TRUE}
}
\description{
load_shp2_by_country
}
