% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_targeting.R
\name{load_targets_by_country}
\alias{load_targets_by_country}
\title{load_targets_by_country}
\usage{
load_targets_by_country(datapath, modelpath, country)
}
\arguments{
\item{datapath}{path to data}

\item{modelpath}{modelpath}

\item{country}{country code}
}
\value{
dataframe with incidence and population by admin unit for a single country
}
\description{
Load possible vaccination targets by country, including incidence and population proportion information
}
