% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_worldpop_by_country.R
\name{load_worldpop_by_country}
\alias{load_worldpop_by_country}
\title{load_worldpop_by_country}
\usage{
load_worldpop_by_country(datapath, country)
}
\arguments{
\item{datapath}{path to data}

\item{country}{country code}
}
\value{
population raster
}
\description{
Load worldpop population raster for a country
}
