% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R,
%   R/utils_process_rawoutput.R
\name{make_eff_table}
\alias{make_eff_table}
\alias{make_eff_table_median}
\title{make_eff_table}
\usage{
make_eff_table(rc)

make_eff_table_median(rc)
}
\arguments{
\item{rc}{target table of one country at one admin level}
}
\value{
a table summarizing target population and ocv efficiency for one country at one admin_level (either admin1 or admin2)

a table summarizing target population and ocv efficiency for one country at one admin_level (either admin1 or admin2)
}
\description{
make a table summarizing (total) target population and ocv efficiency (at country level)

make a table summarizing (total) target population and ocv efficiency (at country level)
}
