% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R,
%   R/utils_process_rawoutput.R
\name{make_eff_table_allISOs}
\alias{make_eff_table_allISOs}
\title{make_eff_table_allISOs}
\usage{
make_eff_table_allISOs(rc)

make_eff_table_allISOs(rc)
}
\arguments{
\item{rc}{target table all countries at one admin level}
}
\value{
a table summarizing target population and ocv efficiency for all countries at one admin_level (either admin1 or admin2)

a table summarizing target population and ocv efficiency for all countries at one admin_level (either admin1 or admin2)
}
\description{
make a table summarizing (total) target population and ocv efficiency, combining all the countries all at once

make a table summarizing (total) target population and ocv efficiency, combining all the countries all at once
}
