% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_rawoutput.R
\name{make_target_table}
\alias{make_target_table}
\title{make_target_table}
\usage{
make_target_table(
  countries,
  thresholds,
  surveillance_scenarios = c("no-estimate", "global-estimate", "district-estimate"),
  output_raw_path = "output_raw/202110gavi-3"
)
}
\arguments{
\item{countries}{vector of ISO country code(s)}

\item{thresholds}{selected thresholds, e.g. all thresholds: thresholds = c(0.001, 2e-04, 1e-04)}

\item{surveillance_scenarios}{selected scenarios, e.g. all scenarios: surveillance_scenarios = c("no-estimate", "global-estimate", "district-estimate")}

\item{output_raw_path}{path to the folders where raw output dataframes are stored}
}
\value{
a combined target_table for selected countries (with true cases number of different scenarios)
}
\description{
combine all raw outputs from selected countries to make a combined target table
}
