% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_skip_countries.R
\name{novacc_campde_transfer}
\alias{novacc_campde_transfer}
\title{novacc_campde_transfer}
\usage{
novacc_campde_transfer(rawoutpath, config)
}
\arguments{
\item{rawoutpath}{path to the raw output folder}

\item{config}{the config with all the parameters}
}
\description{
novacc_campde_transfer
}
