% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{omega_alpha_distribution}
\alias{omega_alpha_distribution}
\title{omega_alpha_distribution}
\usage{
omega_alpha_distribution(
  countries,
  surveillance_scenario_chosen = "global-estimate",
  admins = "both",
  year_chosen = 2022,
  threshold_chosen = 0.001,
  cache,
  ...
)
}
\arguments{
\item{countries}{vector of all the countries included the diagnostic report}

\item{surveillance_scenario_chosen}{scale of surveillance, takes "no-estimate", "global-estimate", or "district-estimate"}

\item{admins}{admin levels to include}

\item{year_chosen}{}

\item{threshold_chosen}{}

\item{cache}{where the filenames and files are saved}
}
\value{
two distribution plots cached
}
\description{
plot the distribution of omega d
}
