% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_incidence_rate_multiplier.R
\name{outbreak_incidence_rate_multiplier}
\alias{outbreak_incidence_rate_multiplier}
\title{outbreak_incidence_rate_multiplier}
\usage{
outbreak_incidence_rate_multiplier(
  datapath,
  modelpath,
  country,
  lambda,
  population_raster,
  output_years,
  use_country_incid_trend,
  use_original_AR = TRUE,
  draw_district_AR = TRUE,
  use_threshold = TRUE,
  smaller_than_1_multiplier_accept = FALSE,
  diagnosis = FALSE,
  outbreak_df_based_scalar = TRUE
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{lambda}{the incidence rate raster data that already has been generated}

\item{population_raster}{the 101-layer population raster that also has already been generated}

\item{output_years}{the output years for simulation}

\item{use_country_incid_trend}{whether use the country-level incidence rate trend}

\item{use_original_AR}{whether directly use the attack rate from the dataset}

\item{draw_district_AR}{whether just use the attack rate specific to the district}

\item{use_threshold}{whether use the threshold incidence rate for the non-outbreak months}

\item{smaller_than_1_multiplier_accept}{whether accept the multipliers that are smaller than 1}

\item{diagnosis}{whether do diagnosis}

\item{outbreak_df_based_scalar}{the multiplier/scalar is solely based on the outbreak dataset}
}
\value{
a function that can take in years and give an outbreak multiplier prediction for a given country
}
\description{
Create and return a function that can give a prediction of the district-based incidence rate in that country
}
