% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct_protect_all_ages_2d}
\alias{pct_protect_all_ages_2d}
\title{pct_protect_all_ages_2d}
\usage{
pct_protect_all_ages_2d(years, proportion_under5, my_trunc_year = 10)
}
\arguments{
\item{years}{number of years since vacc campaign}

\item{proportion_under5}{proportion of population under 5 years}

\item{my_trunc_year}{number of years after which VE is 0, default = 10}
}
\value{
proportion protected for a given time since vaccination and proportion under 5
}
\description{
2 dose VE function for all age groups
}
