% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct_protect_over5_1d}
\alias{pct_protect_over5_1d}
\title{pct_protect_over5_1d}
\usage{
pct_protect_over5_1d(years, my_trunc_year = 10)
}
\arguments{
\item{years}{number of years since vacc campaign}

\item{my_trunc_year}{number of years after which VE is 0, default = 10}
}
\value{
proportion protected for a given time since vaccination and over 5s
}
\description{
1 dose VE function for over 5s
}
