% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{plot_cases}
\alias{plot_cases}
\title{plot_cases}
\usage{
plot_cases(
  cache,
  case_type,
  threshold,
  cumulative_type,
  no_vaccination_surveillance_scenario
)
}
\arguments{
\item{cache}{the cached environment}

\item{case_type}{type of the cases to plot}

\item{threshold}{threshold to plot}

\item{cumulative_type}{either by year (non_cumulative), or cumulated across years, or cumulated within each district}

\item{no_vaccination_surveillance_scenario}{the surveillance scenario for the no vaccination scenario}
}
\value{
cached file names
}
\description{
plot the cases and return the ggplot object
}
