% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{plot_efficiency}
\alias{plot_efficiency}
\title{plot_efficiency}
\usage{
plot_efficiency(cache, compare_type, threshold, cumulative_type)
}
\arguments{
\item{cache}{the cached environment}

\item{compare_type}{type of the comparison}

\item{threshold}{threshold to plot}

\item{cumulative_type}{either by year (non_cumulative) or cumulated across years}
}
\value{
cached file names
}
\description{
plot the efficacy and return the ggplot object
}
