% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{plot_incid}
\alias{plot_incid}
\title{plot_incid}
\usage{
plot_incid(df_target, threshold_chosen, ...)
}
\arguments{
\item{df_target}{input dataset derived from cache$target_table}

\item{threshold_chosen}{one threshold chosen to plot incidence}
}
\value{
faceted plots of incidence rate for one chosen threshold
}
\description{
Plot incidence rate trend for one chosen threshold
}
