% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{plot_time_table}
\alias{plot_time_table}
\title{plot_time_table}
\usage{
plot_time_table(cache)
}
\arguments{
\item{cache}{the cached environment}
}
\value{
the kable object
}
\description{
plot the time table
}
