% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_pop_weighted_incid.R
\name{pop_weighted_admin_mean_incid}
\alias{pop_weighted_admin_mean_incid}
\title{pop_weighted_admin_mean_incid}
\usage{
pop_weighted_admin_mean_incid(
  datapath,
  modelpath,
  incidence_rate_raster,
  pop_raster,
  country,
  admin_shp
)
}
\arguments{
\item{datapath}{path to input data}
}
\description{
pop_weighted_admin_mean_incid
}
