% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{prepare_config}
\alias{prepare_config}
\title{prepare_config}
\usage{
prepare_config(p, configpath)
}
\arguments{
\item{p}{dataframe with config parameters, one config per row}

\item{configpath}{path to config files}
}
\value{
Message returning name of yaml config file
}
\description{
Automate generation of model configuration files from a dataframe of parameters. Generates one config per row in the p dataframe.
}
