% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_montagu_centralburden_template.R
\name{retrieve_montagu_centralburden_template}
\alias{retrieve_montagu_centralburden_template}
\title{retrieve_montagu_centralburden_template}
\usage{
retrieve_montagu_centralburden_template(
  modelpath,
  group_id = "JHU-Lee",
  expectations_all = TRUE
)
}
\arguments{
\item{modelpath}{path to Mongatu files}

\item{group_id}{the id of our modeling group (default = 'JHU-Lee')}

\item{expectations_all}{the expectations that VIMC wants for us (default = TRUE)}
}
\value{
single or multiple CSV data files
}
\description{
Retrieves the CSV file(s) with central burden estimates template from the Montagu API.
}
