% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_montagu_coverage.R
\name{retrieve_montagu_coverage}
\alias{retrieve_montagu_coverage}
\title{retrieve_montagu_coverage}
\usage{
retrieve_montagu_coverage(
  modelpath,
  group_id = "JHU-Lee",
  scenarios_all = TRUE
)
}
\arguments{
\item{modelpath}{path to Mongatu files}

\item{group_id}{the id of our modeling group (default = 'JHU-Lee')}

\item{scenarios_all}{the scenario that we want the coverage data from (default = TRUE)}
}
\value{
single or multiple CSV data files
}
\description{
Retrieves the CSV file(s) with vaccination coverage values for the scenario from the Montagu API.
}
