% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_montagu_population.R
\name{retrieve_montagu_population}
\alias{retrieve_montagu_population}
\title{retrieve_montagu_population}
\usage{
retrieve_montagu_population(
  modelpath,
  touchstone,
  group_id = "JHU-Lee",
  expectations_all = TRUE,
  countries_all = TRUE
)
}
\arguments{
\item{modelpath}{path to Mongatu files}

\item{touchstone}{the unique id of the simulation}

\item{group_id}{the id of our modeling group (default = 'JHU-Lee')}

\item{expectations_all}{the expectations that VIMC wants for us (default = TRUE)}

\item{countries_all}{the countries that we want the total population data from (default = TRUE)}
}
\value{
single or multiple CSV data files
}
\description{
Get total country population from standardized source in Montagu.
}
