% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_country_scenario.R
\name{run_country_scenario}
\alias{run_country_scenario}
\title{run_country_scenario}
\usage{
run_country_scenario(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  nsamples,
  ve_direct = generate_pct_protect_function(),
  indirect_mult = generate_indirect_incidence_mult(),
  secular_trend_mult = generate_flatline_multiplier(),
  clean = TRUE,
  redraw = FALSE,
  ...
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{nsamples}{number of stochastic samples to use}

\item{ve_direct}{vaccine effect function (default: generate_pct_protect_function())}

\item{indirect_mult}{indirect effects function (default: generate_indirect_incidence_mult())}

\item{secular_trend_mult}{function that returns a multiplier for secular incidence trends over time (default: generate_flatline_multiplier())}

\item{clean}{logical that indicates whether existing targeting and model outputs (sus, pop, vacc) should be deleted (default = TRUE)}

\item{redraw}{logical that indicates whether existing incidence raster samples should be redrawn (default = FALSE)}

\item{...}{Optional parameters to pass to [`assign_vaccine_targets()`]. See [`assign_vaccine_targets()`] for defaults.}

\item{mu}{inverse life expectancy for country in years^(-1)}
}
\value{

}
\description{
Run full country scenario
}
