% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_targeting.R
\name{run_targeting_strategy}
\alias{run_targeting_strategy}
\title{run_targeting_strategy}
\usage{
run_targeting_strategy(targets_df, targeting_strat)
}
\arguments{
\item{targets_df}{dataframe with possible targets}

\item{targeting_strat}{Unique string to indicate which targeting strategy should be used (currently supports: incidence, affected_pop)}
}
\value{
dataframe that has been rearranged in order that targeting should occur
}
\description{
Arrange dataframe so that possible target locations are listed in order of priority
}
