% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_update_vac_raster.R
\name{save_vac_raster}
\alias{save_vac_raster}
\title{save_vac_raster}
\usage{
save_vac_raster(
  datapath,
  modelpath,
  country,
  nsamples,
  model_year,
  input_list,
  rawoutpath = NULL,
  clean = NULL,
  scenario = NULL,
  incidence_rate_trend = NULL,
  outbreak_multiplier = NULL,
  vac_incid_threshold = NULL,
  surveillance_scenario = NULL
)
}
\value{
save raster
}
\description{
save_vac_raster
}
