% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{surveillance_add_rc_new_row}
\alias{surveillance_add_rc_new_row}
\title{surveillance_add_rc_new_row}
\usage{
surveillance_add_rc_new_row(
  rc_list,
  ec_list,
  pop,
  model_year,
  sim_start_year,
  sim_end_year,
  shp1,
  shp2,
  nsamples,
  mean_ir_span
)
}
\value{
rc_list
}
\description{
quite similar to the load baseline incidence function, but focused on using the expected cases raster and adding new rows instead
}
