% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_create_expectedCases.R
\name{surveillance_create_expectedCases}
\alias{surveillance_create_expectedCases}
\title{surveillance_create_expectedCases}
\usage{
surveillance_create_expectedCases(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  indirect_mult,
  secular_trend_mult,
  nsamples,
  clean,
  redraw,
  sus_list,
  pop,
  model_year,
  config,
  rc_targeted
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{indirect_mult}{function for calculating indirect effects}

\item{secular_trend_mult}{function for calculating secular trends in mean annual incidence}

\item{nsamples}{number of stochastic samples to use}

\item{redraw}{logical indicate whether to redraw incidence raster samples; pass to [`create_incid_raster()`]}
}
\description{
Create proportion of population vaccinated and total population rasterStacks. Write vaccination raster to file and export population raster
}
