% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_true_confirmation_rate.R
\name{surveillance_true_confirmation_rate}
\alias{surveillance_true_confirmation_rate}
\title{surveillance_true_confirmation_rate}
\usage{
surveillance_true_confirmation_rate(country, admin_level)
}
\arguments{
\item{country}{country code}

\item{admin_level}{admin level}
}
\description{
surveillance_true_confirmation_rate
}
