% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{tab_incid}
\alias{tab_incid}
\title{tab_incid}
\usage{
tab_incid(df_target)
}
\arguments{
\item{df_target}{input dataset derived from cache$target_table}
}
\description{
make a table of median of incidence rate across all runs
}
