% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_diagnostic_functions.R
\name{temporal_incid_trend}
\alias{temporal_incid_trend}
\title{temporal_incid_trend}
\usage{
temporal_incid_trend(
  cache,
  random_seed = 1,
  sample_size = 20,
  iso_code,
  admin_level,
  threshold,
  surv_scenario = c("no-estimate", "global-estimate", "district-estimate"),
  simu_id,
  skip_nonvacc = TRUE
)
}
\description{
make figures of incidence rate trend at the district level
}
