% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{update_novacc_year}
\alias{update_novacc_year}
\title{update_novacc_year}
\usage{
update_novacc_year(
  datapath,
  rc_list,
  model_year,
  surveillance_scenario,
  rc_targeted,
  use_mean_ir,
  mean_ir_span,
  testing_sensitivity
)
}
\value{
rc_list
}
\description{
update the targets in the list if it's not a vaccination campaign year
}
