% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_update_sus_raster.R
\name{update_sus_rasterStack}
\alias{update_sus_rasterStack}
\title{update_sus_rasterStack}
\usage{
update_sus_rasterStack(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  pop,
  ve_direct,
  baseline_year,
  model_year,
  input_list,
  sus_list,
  rc_targeted
)
}
\arguments{
\item{sus_list}{}
}
\value{
sus_list
}
\description{
update one sus rasterStack at a time
}
