% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_update_sus_raster.R
\name{update_sus_rasterStack_optimized}
\alias{update_sus_rasterStack_optimized}
\title{update_sus_rasterStack_optimized}
\usage{
update_sus_rasterStack_optimized(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  pop,
  ve_direct,
  baseline_year,
  model_year,
  last_vac_ef_year = 5,
  rc_targeted
)
}
\arguments{
\item{last_vac_ef_year}{the last year when vac is still effective after received}
}
\description{
update multi-layer sus rasterStack at one time
}
