% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_skip_countries.R
\name{update_table_screening}
\alias{update_table_screening}
\title{update_table_screening}
\usage{
update_table_screening(
  datapath,
  modelpath,
  country,
  scenario,
  threshold,
  vac_start_year,
  vac_end_year,
  rc_list,
  rc_targeted,
  nsamples
)
}
\arguments{
\item{datapath}{path to the input folder}

\item{modelpath}{path to montagu}

\item{country}{country code}

\item{scenario}{}

\item{threshold}{current incidence rate threshold}

\item{vac_start_year}{the year when the vacc is started}

\item{vac_end_year}{the year when the vacc ends}

\item{rc_list}{a list of targeting table to check confirmed_incidence_rate}

\item{rc_targeted}{admin levels to be targeted at}

\item{nsamples}{number of layers}
}
\value{
check
}
\description{
update_table_screening
}
