% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{update_targets_list}
\alias{update_targets_list}
\title{update_targets_list}
\usage{
update_targets_list(
  datapath,
  modelpath,
  country,
  scenario,
  rc_list,
  model_year,
  campaign_cov,
  threshold,
  vac_unconstrained,
  surveillance_scenario,
  vac_interval,
  vac_start_year,
  vac_end_year,
  num_skip_years,
  rc_targeted,
  use_mean_ir,
  mean_ir_span,
  testing_sensitivity
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{}

\item{scenario}{}

\item{rc_list}{}

\item{vac_interval}{country level skipped years}

\item{vac_end_year}{}

\item{num_skip_years}{district level skipped years}

\item{testing_sensitivity}{true sensitivity of testing}
}
\value{
dataframe
}
\description{
update targets in the rc_list
}
