% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_update_vac_raster.R
\name{update_vac_raster}
\alias{update_vac_raster}
\title{update_vac_raster}
\usage{
update_vac_raster(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  rc_list,
  model_year,
  pop,
  input_list,
  no_vacc_year,
  rc_targeted
)
}
\value{
raster
}
\description{
update_vac_raster one at a time
}
