% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vaccine_efficacy_all_groups_all_doses}
\alias{vaccine_efficacy_all_groups_all_doses}
\title{vaccine_efficacy_all_groups_all_doses}
\usage{
vaccine_efficacy_all_groups_all_doses(
  efficacy_one_dose,
  efficacy_two_dose,
  proportion_one_dose
)
}
\arguments{
\item{efficacy_one_dose}{vaccine efficacy for the one-dose group (output of vaccine_efficacy_all_groups_one_dose)}

\item{efficacy_two_dose}{vaccine efficacy for the two-dose group (output of vaccine_efficacy_all_groups_two_dose)}

\item{proportion_one_dose}{the proportion of vaccinated people that received one dose of the vaccine}
}
\value{
the vaccine efficacy for the mixed group that includes people vaccinated with one and two doses of the vaccine
}
\description{
function that takes the vaccine efficacy for the one-dose group and the vaccine efficacy of the two-dose group and 
returns their weighted average. Weighting is according to the proportion people that had one dose/two doses of the vaccine in the population
}
