% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vaccine_efficacy_all_groups_one_dose}
\alias{vaccine_efficacy_all_groups_one_dose}
\title{vaccine_efficacy_all_groups_one_dose}
\usage{
vaccine_efficacy_all_groups_one_dose()
}
\value{
a function returning the weighted average vaccine efficacy for the one-dose campaign
}
\description{
wrapper function that takes the vaccine efficacy functions for the <5 and >5 year olds for the one-dose campaign and returns a new function that returns their weighted average. Weighting is according to the proportion of <5 and >5 year-olds in the population
}
