% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expectedCases.R
\name{create_expectedCases}
\alias{create_expectedCases}
\title{create_expectedCases}
\usage{
create_expectedCases(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  vacc_alloc,
  indirect_mult,
  secular_trend_mult,
  nsamples,
  is_cf,
  redraw
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{vacc_alloc}{object returned from [`allocate_vaccine()`]}

\item{indirect_mult}{function for calculating indirect effects}

\item{secular_trend_mult}{function for calculating secular trends in mean annual incidence}

\item{nsamples}{number of stochastic samples to use}

\item{is_cf}{logical indicating whether to run counterfactual model}

\item{redraw}{logical indicate whether to redraw incidence raster samples; pass to [`create_incid_raster()`]}
}
\value{
dataframe with
}
\description{
Create proportion of population vaccinated and total population rasterStacks. Write vaccination raster to file and export population raster
}
