% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_incid_raster}
\alias{create_incid_raster}
\title{create_incid_raster}
\usage{
create_incid_raster(datapath, country, rawoutpath, nsamples, clean)
}
\arguments{
\item{datapath}{path to data}

\item{country}{country code}

\item{rawoutpath}{path to raw model output files}

\item{nsamples}{numeric, number of layers to sample (must be below 1000)}

\item{clean}{logical that indicates whether existing vacc_files should be deleted}
}
\value{
raster of incidence rate, 30 samples
}
\description{
Generate a raster with the baseline cholera incidence
}
