% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_incid_raster.R
\name{create_incid_raster}
\alias{create_incid_raster}
\title{create_incid_raster}
\usage{
create_incid_raster(modelpath, datapath, country, nsamples, redraw)
}
\arguments{
\item{modelpath}{path to montagu}

\item{datapath}{path to data}

\item{country}{country code}

\item{nsamples}{numeric, number of layers to sample (must be below 1000)}

\item{redraw}{logical that indicates whether existing incid samples should be drawn again}
}
\value{
raster of incidence rate, 30 samples
}
\description{
Generate a raster with the baseline cholera incidence
}
