% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sus_modelInputs.R
\name{create_sus_modelInputs}
\alias{create_sus_modelInputs}
\title{create_sus_modelInputs}
\usage{
create_sus_modelInputs(
  datapath,
  modelpath,
  country,
  scenario,
  rawoutpath,
  vacc_alloc,
  ve_direct,
  clean
)
}
\arguments{
\item{datapath}{path to input data}

\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{scenario}{Unique string that identifies the coverage scenario name}

\item{rawoutpath}{path to raw model output files}

\item{vacc_alloc}{object returned from [`allocate_vaccine()`]}

\item{ve_direct}{vaccine effect function (default: generate_pct_protect_function())}

\item{clean}{logical that indicates whether existing sus_files should be deleted}
}
\description{
Create raster for proportion of population susceptible in each year, after accounting for vaccination
}
