% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_cfr}
\alias{generate_cfr}
\title{generate_cfr}
\usage{
generate_cfr(country)
}
\arguments{
\item{country}{country code}
}
\value{
numeric value of cfr (deaths/cases)
}
\description{
Generate rough estimate of cholera case-fatality ratio based on previous WHO data
}
