% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_flatline_multiplier}
\alias{generate_flatline_multiplier}
\title{generate_flatline_multiplier}
\usage{
generate_flatline_multiplier()
}
\value{

}
\description{
Generate a function that represents projected secular trends in cholera incidence. This multiplier is used to adjust projected cholera incidence in future years.
}
