% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_montagu.R
\name{import_country_population}
\alias{import_country_population}
\title{import_country_population}
\usage{
import_country_population(modelpath, country, redownload = TRUE)
}
\arguments{
\item{modelpath}{path to montagu files}

\item{country}{country code}

\item{redownload}{whether to redownload the file}
}
\description{
Get total country population from standardized source in Montagu. [may be used in future for case-based targeting]
}
