% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_vacc_targeting.R
\name{load_baseline_incidence}
\alias{load_baseline_incidence}
\title{load_baseline_incidence}
\usage{
load_baseline_incidence(
  datapath,
  modelpath,
  country,
  campaign_cov,
  baseline_year,
  first_vacc_year,
  incidence_rate_trend,
  use_country_incid_trend,
  shp0,
  shp1,
  shp2,
  random_seed,
  nsamples,
  redraw
)
}
\arguments{
\item{country}{}
}
\description{
from lambda to incidence rate look-up table to rc_list
}
