% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_custom_shapefile_by_country}
\alias{load_custom_shapefile_by_country}
\title{load_custom_shapefile_by_country}
\usage{
load_custom_shapefile_by_country(country)
}
\arguments{
\item{country}{country code}
}
\value{
shapefile in sf format
}
\description{
Load a custom shapefile for DRC health zones (for the DRC case study) or download or admin0 (simple) level shapefile for a country from GADM, if not already in shapefiles folder, and return sf object
}
